requirements.txt
Django
PyMySQL
django-extensions
django-ratelimit
django-gravatar2
reportlab
sendgrid
tzdata
pyaml
uwsgi
